+++ core-metrics-plus.php
@@ -2,7 +2,7 @@
 /*
 Plugin Name: Core Metrics Plus
 Description: Optimize your WordPress site's Core Web Vitals through smart resource loading and performance enhancements.
-Version: 1.0.0
+Version: 1.1.0
 Author: Carmelyne
 Plugin URI: https://github.com/carmelyne/core-metrics-plus
 */
@@ -40,3 +40,37 @@ function add_fetch_priority() {
 
 // Hook into WordPress footer
 add_action('wp_footer', 'add_fetch_priority');
+
+/**
+ * Add defer attribute to non-essential scripts
+ */
+function cmp_defer_scripts($tag, $handle, $src) {
+    // List of scripts that should not be deferred
+    $no_defer = array(
+        'jquery',              // Core jQuery
+        'jquery-core',         // jQuery core
+        'jquery-migrate'       // jQuery migrate
+    );
+
+    // Don't defer these critical scripts
+    if (in_array($handle, $no_defer)) {
+        return $tag;
+    }
+
+    // Add defer attribute to script tag
+    return str_replace(' src', ' defer src', $tag);
+}
+add_filter('script_loader_tag', 'cmp_defer_scripts', 10, 3);
+
+/**
+ * Optimize script loading order
+ */
+function cmp_optimize_scripts() {
+    if (!is_admin()) {
+        // Move jQuery to footer if it's not needed in header
+        wp_scripts()->add_data('jquery', 'group', 1);
+        wp_scripts()->add_data('jquery-core', 'group', 1);
+        wp_scripts()->add_data('jquery-migrate', 'group', 1);
+    }
+}
+add_action('wp_enqueue_scripts', 'cmp_optimize_scripts', 99);